Change the columns, keys and update the server

+++ src/main/java/com/enbiso/proj/jproject/repository/TaskImportanceRepository.java
@@ -9,6 +9,6 @@ import java.util.List;
 /**
  * Spring Data JPA repository for the TaskImportance entity.
  */
-public interface TaskImportanceRepository extends JpaRepository<TaskImportance,Long> {
+public interface TaskImportanceRepository extends JpaRepository<TaskImportance,Integer> {
 
 }
